Fading between screens working but being a big poopoo (Try write an asnyc function to render all the screens then switch a boolean which loads the first screen)
Car can drive now :D

Search Algo implemented to track path to specific position, need to get sprite to move accordingly now, character can move following the algo, need to add a function to ensure that the sprite starts in the centre of a tile so that the tracking is accurate.

(THIS IS NOT HELPFUL YET LOL)Starting input processors for menu, menu started

Started on flags, Started working on points, Self Resolve working working on making sprites appear, house keeping, flag appear and points added, flags disappear on click now, more housekeeping and documentation, Flag random spawning and points, implemented input listener

Menu & shop class created, properly pausing when menu button in, properly pausing and menu navigation set up, TODO add content and purchasing. MENU AND DAY END SORTA BUT GOTTA DESTROY FIXTURES AT DAY END, ITS FIXED NOW!!! ALL THATS LEFT IS TO ADD THE SHOP AND SAVE AND WE DONE BOIIISS

SHOP FUNCTIONAL, FINE TUNING NEEDED, DUPLICATE CODE FOR CLASS SECTION TO INFRASTRUCTURE TAB, INFRASTRUCTURE TAB FINISHED (sorta), housekeeping, Added a bunch of TODO's so do those and also make sure things are clear in write up and video

TitleScreen stuff finished, Shop stuff next, shop stuff done, need to show error message

+++ core/src/com/mygdx/game/Screens/TitleScreen.java
@@ -61,7 +61,7 @@ public class TitleScreen implements Screen {
     private TextButton start;
     private TextButton delete;
 
-    public TitleScreen(Main game) {
+    public TitleScreen(final Main game) {
         this.myGame = game;
         this.world = new World(new Vector2(0, 0), true);
         world.setContactListener(new WorldContactListener(game));
@@ -97,7 +97,11 @@ public class TitleScreen implements Screen {
         delete.addListener(new com.badlogic.gdx.scenes.scene2d.InputListener(){
             @Override
             public boolean touchDown(InputEvent event, float x, float y, int pointer, int button) {
-                myGame.deleteSave();
+                if(game.getPref().getBoolean(Con.FILE_EXISTS)){
+                    myGame.deleteSave();
+                }
+                else
+                    myGame.showError();
                 return false;
             }
         });
@@ -131,6 +135,9 @@ public class TitleScreen implements Screen {
     @Override
     public void render(float delta) {
         update(delta);
+        if(myGame.getErrorLabel().getColor().a >0) {
+            myGame.fadeError(delta);
+        }
         //clear screen
         Gdx.gl.glClearColor(0, 1 , 1 ,1);
         Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
